robot_sync_app/requirements.txt
pyyaml>=6.0
pyserial>=3.5
boto3>=1.34.0
nvidia-riva-client>=2.19.1
